requirements.txt
streamlit
google-generativeai
python-dotenv
PyPDF2
flask
openai
python-dotenv
pdfplumber